requirements.txt
flask
transformers
torch
openai
boto3
psycopg2-binary
sqlalchemy
matplotlib
python-dotenv